Metal to separate module
