+++ src/material/mod.rs
@@ -1,6 +1,7 @@
 use std::fmt::Debug;
 
 pub use lambertian::*;
+pub use metal::*;
 
 use crate::hittable::Hit;
 use crate::random;
@@ -9,42 +10,12 @@ use crate::texture::{Color, Texture};
 use crate::vec::V3;
 
 pub mod lambertian;
+pub mod metal;
 
 pub trait Material: Debug + Sync + Send {
     fn scatter(&self, ray: &Ray, hit: &Hit) -> Option<Ray> { None }
     fn emmit(&self, hit: &Hit) -> Color { Color(V3::zeros()) }
 }
-
-
-#[derive(PartialEq, Copy, Clone, Debug)]
-pub struct Metal {
-    albedo: V3,
-    fuzz: f64,
-}
-
-impl Metal {
-    pub fn new(albedo: V3) -> Metal { Metal { albedo, fuzz: 0.0 } }
-    pub fn new_fuzzed(albedo: V3, fuzz_factor: f64) -> Metal {
-        Metal { albedo, fuzz: if fuzz_factor < 1.0 { fuzz_factor } else { 1.0 } }
-    }
-
-    fn fuzz(self, vector: V3) -> V3 {
-        self.fuzz * random::rand_in_unit_sphere() + vector
-    }
-}
-
-impl Material for Metal {
-    fn scatter(&self, ray: &Ray, &hit: &Hit) -> Option<Ray> {
-        let unit_direction = ray.direction.unit();
-        let reflected = unit_direction.reflect(hit.normal);
-        if reflected.dot(hit.normal) > 0.0 {
-            Some(ray.produce(hit.point, self.fuzz(reflected), self.albedo))
-        } else {
-            None
-        }
-    }
-}
-
 
 #[derive(PartialEq, Copy, Clone, Debug)]
 pub struct Dielectric {
